+++ deploy/sftp.sh
@@ -1,8 +1,13 @@
 #!/bin/sh
 set -e
 
+if test "$1" = "--pull"; then
+  docker pull nswebfrog/sftp:latest
+  exit
+fi
+
 docker rm -f sftp || true
-docker run -d --pull=always \
+docker run -d \
   --name=sftp \
   --restart=always \
   --log-opt max-size=10m \
